Rework algorithm to apply edit distance to only overlapping insert subsets
In the first version, I simply did an iterative approach (O(n+m)) that
simply "merged" the two operations together. I accounted for the simple
cases, which were all order dependent (e.g. you want removes to only
remove retains). However, when dealing with merging two subsets of ops
that were only inserts, there were duplication of insert ops in both
op1' and op2'.

In the 2nd version, I attempted to tweak the existing algorithm by
incorporating the edit distance. This had an overall runtime of
O(n*m*(n+m)), which is due to the DP nature (n*m) but in each
iteration we had to clone the previous ops' array.

However, there was the major issue where I was working with the
`CompressedOperation`, which meant that the implementation must hold
much more state to keep track of the op's "source" (i.e. whether it came
from the local client or from the server) due to wanting to have
consistent ordering as well as knowing which retains were removed. The
overall implementation was way too complex so I decided to scrap it.

In the 3rd version, I introduced `SingleOperation`, which is the
smallest unit of operation possible (e.g. a single char insert) so that
we can store additional metadata (the source, the relative positioning,
etc). This helped immensely in reducing the complexity and everything
_looked_ good at first as unit tests were passing. My implementation
applied the edit distance to the **whole** op sets. That meant that in
order to guarantee order, in each DP iteration I had to find where to
insert each op. This would later surface as an issue.

During the integration tests, however, there was an issue where we
couldn't guarantee the relative order when it comes to subsets of insert
ops. After some thought, I realized I was applying the edit distance
completely wrong. Instead of applying the edit distance to the whole ops
set, I should have only applied it to overlapping insert ops.

In the 4th and current version, it became a hybrid of the 1st version
(iterate through the ops) but apply the edit distance when encountered
overlapping insert ops.